Create AddressQuestionTest
