+++ universal-application-tool-0.0.1/test/services/applicant/question/AddressQuestionTest.java
@@ -1,0 +1,91 @@
+package services.applicant.question;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.google.common.collect.ImmutableMap;
+import java.util.Locale;
+import models.Applicant;
+import org.junit.Before;
+import org.junit.Test;
+import services.Path;
+import services.applicant.ApplicantData;
+import services.applicant.ValidationErrorMessage;
+import services.question.AddressQuestionDefinition;
+
+public class AddressQuestionTest {
+
+  private static final AddressQuestionDefinition addressQuestionDefinition =
+      new AddressQuestionDefinition(
+          1L,
+          "question name",
+          Path.create("applicant.my.path.name"),
+          "description",
+          ImmutableMap.of(Locale.US, "question?"),
+          ImmutableMap.of(Locale.US, "help text"));
+
+  private Applicant applicant;
+  private ApplicantData applicantData;
+
+  @Before
+  public void setUp() {
+    applicant = new Applicant();
+    applicantData = applicant.getApplicantData();
+  }
+
+  @Test
+  public void addressQuestion_withEmptyApplicantData() {
+    ApplicantQuestion applicantQuestion =
+        new ApplicantQuestion(addressQuestionDefinition, applicantData);
+
+    assertThat(applicantQuestion.getAddressQuestion()).isInstanceOf(AddressQuestion.class);
+    assertThat(applicantQuestion.getQuestionText()).isEqualTo("question?");
+    assertThat(applicantQuestion.hasErrors()).isFalse();
+  }
+
+  @Test
+  public void addressQuestion_withInvalidApplicantData() {
+    applicantData.putString(addressQuestionDefinition.getStreetPath(), "");
+    applicantData.putString(addressQuestionDefinition.getCityPath(), "");
+    applicantData.putString(addressQuestionDefinition.getStatePath(), "");
+    applicantData.putString(addressQuestionDefinition.getZipPath(), "");
+
+    ApplicantQuestion applicantQuestion =
+        new ApplicantQuestion(addressQuestionDefinition, applicantData);
+    AddressQuestion addressQuestion = applicantQuestion.getAddressQuestion();
+
+    assertThat(applicantQuestion.hasErrors()).isTrue();
+    assertThat(addressQuestion.getStreetErrors())
+        .contains(ValidationErrorMessage.create("Street is required."));
+    assertThat(addressQuestion.getCityErrors())
+        .contains(ValidationErrorMessage.create("City is required."));
+    assertThat(addressQuestion.getStateErrors())
+        .contains(ValidationErrorMessage.create("State is required."));
+    assertThat(addressQuestion.getZipErrors())
+        .contains(ValidationErrorMessage.create("Zip code is required."));
+
+    applicantData.putString(addressQuestionDefinition.getZipPath(), "not a zip code");
+    addressQuestion =
+        new ApplicantQuestion(addressQuestionDefinition, applicantData).getAddressQuestion();
+
+    assertThat(addressQuestion.getZipErrors())
+        .contains(ValidationErrorMessage.create("Invalid zip code."));
+  }
+
+  @Test
+  public void addressQuestion_withValidApplicantData() {
+    applicantData.putString(addressQuestionDefinition.getStreetPath(), "85 Pike St");
+    applicantData.putString(addressQuestionDefinition.getCityPath(), "Seattle");
+    applicantData.putString(addressQuestionDefinition.getStatePath(), "WA");
+    applicantData.putString(addressQuestionDefinition.getZipPath(), "98101");
+
+    ApplicantQuestion applicantQuestion =
+        new ApplicantQuestion(addressQuestionDefinition, applicantData);
+    AddressQuestion addressQuestion = applicantQuestion.getAddressQuestion();
+
+    assertThat(applicantQuestion.hasErrors()).isFalse();
+    assertThat(addressQuestion.getStreetValue().get()).isEqualTo("85 Pike St");
+    assertThat(addressQuestion.getCityValue().get()).isEqualTo("Seattle");
+    assertThat(addressQuestion.getStateValue().get()).isEqualTo("WA");
+    assertThat(addressQuestion.getZipValue().get()).isEqualTo("98101");
+  }
+}
