+++ universal-application-tool-0.0.1/test/services/applicant/question/ApplicantQuestionTest.java
@@ -9,13 +9,10 @@ import java.util.EnumSet;
 import java.util.Locale;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
-import models.Applicant;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import services.Path;
 import services.applicant.ApplicantData;
-import services.applicant.ValidationErrorMessage;
 import services.question.AddressQuestionDefinition;
 import services.question.DropdownQuestionDefinition;
 import services.question.NameQuestionDefinition;
@@ -78,15 +75,6 @@ public class ApplicantQuestionTest {
           ImmutableMap.of(Locale.US, "question?"),
           ImmutableMap.of(Locale.US, "help text"));
 
-  private Applicant applicant;
-  private ApplicantData applicantData;
-
-  @Before
-  public void setUp() {
-    applicant = new Applicant();
-    applicantData = applicant.getApplicantData();
-  }
-
   // TODO(https://github.com/seattle-uat/civiform/issues/405): Change this to just use
   // @Parameters(source = QuestionType.class) once RepeatedQuestionDefinition exists.
   @Test
@@ -101,63 +89,6 @@ public class ApplicantQuestionTest {
 
   private EnumSet<QuestionType> types() {
     return EnumSet.complementOf(EnumSet.of(QuestionType.REPEATER));
-  }
-
-  @Test
-  public void addressQuestion_withEmptyApplicantData() {
-    ApplicantQuestion applicantQuestion =
-        new ApplicantQuestion(addressQuestionDefinition, applicantData);
-
-    assertThat(applicantQuestion.getAddressQuestion()).isInstanceOf(AddressQuestion.class);
-    assertThat(applicantQuestion.getQuestionText()).isEqualTo("question?");
-    assertThat(applicantQuestion.hasErrors()).isFalse();
-  }
-
-  @Test
-  public void addressQuestion_withInvalidApplicantData() {
-    applicantData.putString(addressQuestionDefinition.getStreetPath(), "");
-    applicantData.putString(addressQuestionDefinition.getCityPath(), "");
-    applicantData.putString(addressQuestionDefinition.getStatePath(), "");
-    applicantData.putString(addressQuestionDefinition.getZipPath(), "");
-
-    ApplicantQuestion applicantQuestion =
-        new ApplicantQuestion(addressQuestionDefinition, applicantData);
-    AddressQuestion addressQuestion = applicantQuestion.getAddressQuestion();
-
-    assertThat(applicantQuestion.hasErrors()).isTrue();
-    assertThat(addressQuestion.getStreetErrors())
-        .contains(ValidationErrorMessage.create("Street is required."));
-    assertThat(addressQuestion.getCityErrors())
-        .contains(ValidationErrorMessage.create("City is required."));
-    assertThat(addressQuestion.getStateErrors())
-        .contains(ValidationErrorMessage.create("State is required."));
-    assertThat(addressQuestion.getZipErrors())
-        .contains(ValidationErrorMessage.create("Zip code is required."));
-
-    applicantData.putString(addressQuestionDefinition.getZipPath(), "not a zip code");
-    addressQuestion =
-        new ApplicantQuestion(addressQuestionDefinition, applicantData).getAddressQuestion();
-
-    assertThat(addressQuestion.getZipErrors())
-        .contains(ValidationErrorMessage.create("Invalid zip code."));
-  }
-
-  @Test
-  public void addressQuestion_withValidApplicantData() {
-    applicantData.putString(addressQuestionDefinition.getStreetPath(), "85 Pike St");
-    applicantData.putString(addressQuestionDefinition.getCityPath(), "Seattle");
-    applicantData.putString(addressQuestionDefinition.getStatePath(), "WA");
-    applicantData.putString(addressQuestionDefinition.getZipPath(), "98101");
-
-    ApplicantQuestion applicantQuestion =
-        new ApplicantQuestion(addressQuestionDefinition, applicantData);
-    AddressQuestion addressQuestion = applicantQuestion.getAddressQuestion();
-
-    assertThat(applicantQuestion.hasErrors()).isFalse();
-    assertThat(addressQuestion.getStreetValue().get()).isEqualTo("85 Pike St");
-    assertThat(addressQuestion.getCityValue().get()).isEqualTo("Seattle");
-    assertThat(addressQuestion.getStateValue().get()).isEqualTo("WA");
-    assertThat(addressQuestion.getZipValue().get()).isEqualTo("98101");
   }
 
   @Test
